refactor(forks/lstar): split spec.py into per-concern mixins
Break the single large LstarSpec class into one file per concern (state
transition, fork choice, block production, signatures, aggregation, timeline,
validator duties), assembled back into LstarSpec via mixins over a shared
typing contract (LstarSpecContract).

Pure structural move: every method body is byte-for-byte identical, with no
change to any external call site or type annotation. The only edits are
relocating the chain-match static helper to a module-level function and
moving the LstarStore alias into the contract module (still re-exported from
spec.py).
